[FE] 웹팩 설정 파일을 분리하고 prod 모드일 때 babel-loader를 사용하도록 수정한다.  (#641)
