+++ frontend/.babelrc
@@ -1,0 +1,8 @@
+{
+  "presets": [
+    "@babel/preset-env",
+    "@babel/preset-typescript",
+    ["@babel/preset-react", { "runtime": "automatic", "importSource": "@emotion/react" }]
+  ],
+  "plugins": ["@emotion"]
+}
