+++ frontend/package.json
@@ -4,12 +4,12 @@
   "main": "index.tsx",
   "license": "MIT",
   "scripts": {
-    "build": "cross-env NODE_ENV=production API_ENV=production webpack --mode=production",
-    "build:dev": "cross-env NODE_ENV=production API_ENV=development webpack --mode=development",
-    "build:analyze": "cross-env NODE_ENV=production API_ENV=production BUNDLE_ANALYZE=on webpack --mode=production",
-    "prod": "cross-env NODE_ENV=production API_ENV=production webpack serve --mode=production --open",
-    "dev": "cross-env NODE_ENV=development API_ENV=development webpack serve --mode=development --open",
-    "msw": "cross-env NODE_ENV=development API_ENV=msw webpack serve --mode=development --open",
+    "build": "cross-env NODE_ENV=production API_ENV=production webpack --mode=production --config webpack.prod.js",
+    "build:dev": "cross-env NODE_ENV=production API_ENV=development webpack --mode=development --config webpack.dev.js",
+    "build:analyze": "cross-env NODE_ENV=production API_ENV=production BUNDLE_ANALYZE=on webpack --mode=production --config webpack.prod.js",
+    "prod": "cross-env NODE_ENV=production API_ENV=production webpack serve --mode=production --open --config webpack.prod.js",
+    "dev": "cross-env NODE_ENV=development API_ENV=development webpack serve --mode=development --open --config webpack.dev.js",
+    "msw": "cross-env NODE_ENV=development API_ENV=msw webpack serve --mode=development --open --config webpack.dev.js",
     "init-msw": "msw init public/ --save",
     "storybook": "storybook dev -p 6006",
     "build-storybook": "storybook build",
@@ -20,11 +20,18 @@
     "lint:styled": "stylelint --cache --fix ./*"
   },
   "dependencies": {
+    "@babel/cli": "^7.25.6",
+    "@babel/core": "^7.25.2",
+    "@babel/preset-env": "^7.25.4",
+    "@babel/preset-react": "^7.24.7",
+    "@babel/preset-typescript": "^7.24.7",
+    "@emotion/babel-plugin": "^11.12.0",
     "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.5",
     "@sentry/react": "^8.22.0",
     "@tanstack/react-query": "^5.51.15",
     "@uiw/react-markdown-preview": "^5.1.2",
+    "babel-loader": "^9.2.1",
     "init": "^0.1.2",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
@@ -59,6 +66,7 @@
     "@webpack-cli/generators": "^3.0.7",
     "chromatic": "^11.5.6",
     "concurrently": "^8.2.2",
+    "copy-webpack-plugin": "^12.0.2",
     "cross-env": "^7.0.3",
     "css-loader": "^7.1.2",
     "dotenv": "^16.4.5",
