+++ frontend/src/components/NewChecklist/NewChecklistTemplate.tsx
@@ -2,10 +2,11 @@ import styled from '@emotion/styled';
 
 import Layout from '@/components/_common/layout/Layout';
 import { useTabContext } from '@/components/_common/Tabs/TabContext';
-import ChecklistQuestion from '@/components/NewChecklist/ChecklistQuestion/ChecklistQuestion';
+import ChecklistQuestionItem from '@/components/NewChecklist/ChecklistQuestion/ChecklistQuestion';
 import useChecklistStore from '@/store/useChecklistStore';
 import { flexColumn } from '@/styles/common';
 import theme from '@/styles/theme';
+import { ChecklistQuestion } from '@/types/checklist';
 
 const NewChecklistTemplate = () => {
   const { currentTabId } = useTabContext();
@@ -17,7 +18,7 @@ const NewChecklistTemplate = () => {
     <Layout bgColor={theme.palette.background} withHeader withTab>
       <S.ContentBox>
         {questions?.questions.map((question: ChecklistQuestion) => (
-          <ChecklistQuestion key={`${currentTabId}-${question.questionId}`} question={question} />
+          <ChecklistQuestionItem key={`${currentTabId}-${question.questionId}`} question={question} />
         ))}
       </S.ContentBox>
     </Layout>
